Провел рефакторинг
