+++ js/main.js
@@ -1,9 +1,8 @@
-import './img-upload.js';
 import {addThumbnails,} from './add-thumbnails.js';
 import {getData} from './api.js';
 import {initializeGallery} from './big-photo-mode.js';
 import {showErrorMessage} from './util.js';
-
+import {initSendForm} from './send-form.js';
 
 getData()
   .then((photos) => {
@@ -14,3 +13,4 @@ getData()
     showErrorMessage(error.message);
   });
 
+initSendForm();
